pyproject.toml
[tool.poetry]
name = "componentsemantics"
version = "0.1.0"
description = ""
authors = ["Cezar  Sas"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
backcall = "0.2.0"
beautifulsoup4 = "4.9.1"
bitarray = "^2.3.4"
bkcharts = "0.2"
bokeh = "^2.4.0"
boto = "2.49.0"
Bottleneck = "1.3.2"
brotlipy = "0.7.0"
cairocffi = "1.1.0"
catalogue = "1.0.0"
certifi = "2020.6.20"
cffi = "1.14.0"
chardet = "3.0.4"
click = "7.1.2"
cloudpickle = "^2.0.0"
Cycler = "0.10.0"
cymem = "2.0.3"
Cython = "^0.29.24"
cytoolz = "0.10.1"
dask = "^2021.9.1"
DataShape = "0.5.2"
decorator = "4.4.2"
defusedxml = "0.6.0"
diff-match-patch = "^20200713"
distributed = "^2021.9.1"
docutils = "0.16"
entrypoints = "0.3"
et-xmlfile = "1.0.1"
fastcache = "1.1.0"
fastrlock = "0.5"
filelock = "3.0.12"
flake8 = "3.8.3"
Flask = "1.1.2"
fsspec = "0.7.4"
ftfy = "5.8"
future = "0.18.2"
glob2 = "0.7"
HeapDict = "1.0.1"
holoviews = "1.13.4"
html5lib = "^1.1"
idna = "^3.2"
imageio = "^2.9.0"
imagesize = "1.2.0"
importlib-metadata = "^4.8.1"
infomap = "1.1.4"
intervaltree = "^3.1.0"
ipykernel = "^6.4.1"
ipython = "^7.28.0"
ipython_genutils = "0.2.0"
ipywidgets = "7.5.1"
isort = "4.3.21"
itsdangerous = "1.1.0"
jeepney = "^0.7.1"
Jinja2 = "2.11.2"
jmespath = "0.10.0"
joblib = "^1.0.1"
jupyter = "1.0.0"
jupyter-client = "^7.0.3"
jupyter-console = "6.1.0"
jupyter-core = "4.6.3"
jupyterlab = "^3.1.13"
jupyterlab-server = "^2.8.2"
keyring = "^23.2.1"
lxml = "^4.6.3"
Markdown = "3.3.2"
MarkupSafe = "1.1.1"
matplotlib = "^3.4.3"
networkx = "^2.6.3"
nltk = "^3.6.3"
nose = "1.3.7"
numpydoc = "^1.1.0"
olefile = "0.46"
openpyxl = "^3.0.9"
packaging = "20.4"
pandas = "^1.3.3"
pathtools = "0.1.2"
patsy = "0.5.1"
pep8 = "1.7.1"
pexpect = "4.8.0"
pickleshare = "0.7.5"
Pillow = "^8.3.2"
pkginfo = "1.5.0.1"
plac = "1.1.3"
pluggy = "0.13.1"
ply = "3.11"
py = "^1.10.0"
pycodestyle = "2.6.0"
pycosat = "0.6.3"
pycparser = "^2.20"
pyct = "0.4.6"
pycurl = "7.43.0.5"
pydocstyle = "^6.1.1"
pyflakes = "2.2.0"
Pygments = "2.6.1"
pylint = "^2.11.1"
pyOpenSSL = "^20.0.1"
pyparsing = "2.4.7"
pyrsistent = "0.16.0"
PySocks = "1.7.1"
pytest = "5.4.3"
python-dateutil = "2.8.1"
python-igraph = "0.8.2"
pytokenizations = "0.6.0"
pytz = "2020.1"
pyviz-comms = "0.7.6"
PyYAML = "5.3.1"
regex = "^2021.9.24"
requests = "^2.26.0"
rope = "0.17.0"
"ruamel.yaml" = "0.15.87"
s3transfer = "0.3.3"
sacremoses = "0.0.43"
scikit-image = "0.16.2"
scikit-learn = "^1.0"
scipy = "^1.7.1"
seaborn = "0.11.0"
SecretStorage = "^3.3.1"
Send2Trash = "1.5.0"
sentencepiece = "0.1.91"
simplegeneric = "0.8.1"
snowballstemmer = "2.0.0"
sortedcollections = "1.2.1"
sortedcontainers = "2.2.2"
sourcy = "0.2.2"
spacy = "2.3.2"
spacy-transformers = "0.6.2"
Sphinx = "^4.2.0"
SQLAlchemy = "^1.4.25"
tokenizers = "0.5.2"
toml = "^0.10.2"
toolz = "0.10.0"
torch = "1.6.0"
torchcontrib = "0.0.2"
torchvision = "0.7.0"
tornado = "^6.1"
tqdm = "^4.62.3"
traitlets = "4.3.3"
transformers = "2.8.0"
tree-sitter = "0.1.1"
typing-extensions = "^3.10.0"
ujson = "1.35"
umap-learn = "0.4.6"
unicodecsv = "0.14.1"
urllib3 = "1.25.9"
wasabi = "0.8.0"
watchdog = "^2.1.5"
wcwidth = "^0.2.5"
webencodings = "0.5.1"
Werkzeug = "1.0.1"
widgetsnbextension = "3.5.1"
wrapt = "1.11.2"
wurlitzer = "^3.0.2"
leidenalg = "0.8.2"
xarray = "0.16.1"
xlrd = "1.2.0"
XlsxWriter = "1.2.9"
xlwt = "1.3.0"
yapf = "^0.31.0"
zict = "2.0.0"
zipp = "3.1.0"
"zope.event" = "4.4"
"zope.interface" = "4.7.1"
textblob = "^0.17.1"
fasttext = "^0.9.2"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
jupyterlab = "^3.2.9"
notebook = "6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
